Correct syntax for pth file
PTH files are read line by line, with a path or, if starting with an
import, will execute the line.
So we use this trick to make the try/import/except/pass in a `exec`
call.

The `except ImportError` is useful when upgrading mixt version, because
the package is first uninstalled and in this case the import fail.