Manage html5 boolean attributes
In html 5, a boolean attribute is considered False if not set, and True
if set without value, or if a value is given, it must be an empty string
or the name of the attribute itself.

In addition, we allow True/False or a string which is, lowercased,
"true" or "false".