Use python typing instead of a dict to handle attr types
This is done by using a class `Attrs` instead of a dict named
`__attrs__`.

Only attributes of this class with a defined type will be used as real
attributes.
These attributes must be simple to be used as a cast function.
A specific type is `Choices`, and must be used when the value must be in
a list of valid values.

Also, now that it's a class with attributes, and not a dict, attribute
names must be valid identifier. So:
- use `_` instead of the `-` of an html attribute
- use `__` instead of the `:` of an html attribute
- prefix with `_` if the name is not a valid identifier (like `class` or
`for`, for example)