Allow pylint to work with multilines html mixt files
The ``EncodingChecker`` pylint linter try to encode files line by line,
but the ``mixt`` codec cannot do that as it have to find all closing
tags, and they may be not on the same line.

So for ``mixt`` encoded files, we tell this linter to no trying to
decode it and assume its unicode.

It's done via ``patchy``, a controversial way to monkey-patch, but:
1. it works.
2. it is clearly readable.